Move cli/cmd/test_helpers.go to pkg/profiles/test_helper.go, address review comments
