+++ web/srv/handlers_test.go
@@ -9,10 +9,10 @@ import (
 
 	"github.com/ghodss/yaml"
 	"github.com/julienschmidt/httprouter"
-	helpers "github.com/linkerd/linkerd2/cli/cmd"
 	"github.com/linkerd/linkerd2/controller/api/public"
 	"github.com/linkerd/linkerd2/controller/gen/apis/serviceprofile/v1alpha1"
 	pb "github.com/linkerd/linkerd2/controller/gen/public"
+	helpers "github.com/linkerd/linkerd2/pkg/profiles"
 )
 
 func TestHandleIndex(t *testing.T) {
@@ -109,7 +109,7 @@ func TestHandleConfigDownload(t *testing.T) {
 		t.Fatalf("Error parsing service profile: %v", err)
 	}
 
-	expectedServiceProfile := helpers.GenServiceProfile("authors", "booksns")
+	expectedServiceProfile := helpers.GenServiceProfile("authors", "booksns", "linkerd")
 
 	err = helpers.ServiceProfileYamlEquals(serviceProfile, expectedServiceProfile)
 	if err != nil {
